+++ src/DurableTask/TaskOrchestrationContext.cs
@@ -37,7 +37,7 @@ public abstract class TaskOrchestrationContext
     /// The current orchestration time is stored in the orchestration history and this API will
     /// return the same value each time it is called from a particular point in the orchestration's
     /// execution. It is a deterministic, replay-safe replacement for existing .NET APIs for getting
-    /// the curren time, such as <see cref="DateTime.UtcNow"/> and <see cref="DateTimeOffset.UtcNow"/>.
+    /// the current time, such as <see cref="DateTime.UtcNow"/> and <see cref="DateTimeOffset.UtcNow"/>.
     /// </remarks>
     public abstract DateTime CurrentUtcDateTime { get; }