+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## v0.5.0-beta
 
+- Adds `PurgeInstancesMetadataAsync` and `PurgeInstancesAsync` support and implementation to `DurableTaskGrpcClient`
+
 ### Breaking changes
 
 - Added new abstract property `TaskOrchestrationContext.ParentInstance`.
+- Added new abstract method `DurableTaskClient.PurgeInstancesAsync`.
 
 ## v0.4.1-beta